Separate file for Snapshot
